Project final edits with App/dtplc/pyproject.toml
[tool.poetry]
name = "ubi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
onnx = "^1.15.0"
opencv-python = "^4.8.1.78"
pillow = "^10.1.0"
onnxruntime = "^1.16.3"
numpy = "^1.26.2"
scipy = "^1.11.4"
open-clip-torch = "^2.23.0"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
